+++ src/main/java/com/mattaeng/mattaengapi/common/config/P6SpyConfig.java
@@ -1,4 +1,4 @@
-package common.config;
+package com.mattaeng.mattaengapi.common.config;
 
 import com.p6spy.engine.logging.Category;
 import com.p6spy.engine.spy.P6SpyOptions;
